Complete remake of using JSON format
The ReadElement and WriteElement files have been remade to be more robust, easy to use, format mistake resistant and especially support version compatibility.
The success and fail values are now stored only in main part of Word and Picture, no longer in the other part.
Redesigned version list to RecyclerViews.